requirements.txt
flask>=3.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
requests>=2.31.0
urllib3>=2.0.0
